backend/requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cmudict==1.0.32
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
contourpy==1.3.2
coverage==7.9.0
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
deepface==0.0.93
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
easyocr==1.7.2
edfio==0.4.9
EDFlib-Python==1.0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
gTTS==2.5.4
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaggle==1.7.4.5
kagglehub==0.3.12
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
memory-profiler==0.61.0
ml_dtypes==0.5.1
mne==1.9.0
mne-bids==0.16.0
moabb==1.2.0
mpmath==1.3.0
mtcnn==1.0.0
murmurhash==1.0.13
mysql-connector-python==9.3.0
namex==0.1.0
narwhals==1.42.0
networkx==3.5
ninja==1.13.0
nltk==3.8.1
numpy==2.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.3.1
pillow==11.2.1
platformdirs==4.3.8
playsound==1.3.0
plotly==6.2.0
pooch==1.8.2
preshed==3.0.10
protobuf==4.25.7
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pyphen==0.17.2
pypiwin32==223
pyriemann==0.7
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pyttsx3==2.99
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
rpds-py==0.25.1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.14.1
seaborn==0.12.2
sentencepiece==0.2.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sounddevice==0.5.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tesseract==0.1.3
text-unidecode==1.3
textblob==0.19.0
textstat==0.7.3
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.180
ultralytics-thop==2.0.15
urllib3==1.26.20
vaderSentiment==3.3.2
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
zipp==3.23.0
